Add minutes from NYC
I needed to refer to these and thought it might be good to get them out
of the Google doc.

I have not edited these other than to remove the cruft at the top and
bottom.  WYGIWYG.  Once up, I suggest that people can open pull requests
if they find something that is wrong (as in, not a faithful rendition of
what happened during the meeting; clearly, some of the stuff that was
said will prove to be wrong).